refactor: move the platform-owned template tag libraries out of the edition
The platform's shared UI (the cotton component library, the auth templates
and the error pages) loaded template tag libraries that lived in
n23/core/templatetags/. `{% load %}` resolves by library NAME across every
installed app, and a name may only be provided once, so these move rather
than get copied — and the ~215 edition `{% load custom_tags %}` sites need
no change.

Moved to gyrinx/site/templatetags/ unchanged, since none of them import
edition code:

  badge_tags        renders gyrinx.badges supporter badges
  allauth_bootstrap serves the auth templates that moved in #2126
  form_tags         serves gyrinx/templates/cotton/form/ only, imports
                    nothing but django
  group_tags        django.contrib.auth Group membership

Left in the edition:

  color_tags        house icons (n23.content.svg) and gang list theme
                    colours; used only by n23 templates

custom_tags was NOT entirely generic, so a wholesale move would have put
`from n23.content.models import ContentPageRef` at the top of a gyrinx/
module — the exact dependency this work exists to remove. It is split
instead. The generic ~90% becomes gyrinx/site/templatetags/platform_tags.py
with no edition imports; the six tags that genuinely need the edition (ref,
get_skill, credits, fullurl, is_campaign_admin, pack_name) stay in
custom_tags, which re-exports the platform library so edition templates keep
getting the union under one name. Platform templates load platform_tags
directly.

Also: gyrinx/templates/404.html and errors/error.html now extend the
platform base added in #2126 rather than the edition's core/layouts/base.html.

gyrinx/templates/cotton/back.html and cotton/list.html reference
core/includes/back.html and core/includes/rule.html only inside
{% comment %} blocks, as documentation of real call sites — they are not
live includes, so nothing to resolve there.
