+++ gyrinx/site/templatetags/badge_tags.py
