+++ gyrinx/site/templatetags/platform_tags.py
@@ -1,0 +1,699 @@
+"""The platform's general-purpose template tags.
+
+Everything here is edition-agnostic: nav-state helpers, query-string editing,
+small value formatters, HTML sanitisation and the return-URL machinery. Nothing
+in this module may import from an edition package (``n23.*``, ``n26.*``) — that
+is the whole reason it is not still part of the edition's ``custom_tags``.
+
+``n23.core.templatetags.custom_tags`` re-exports everything registered here, so
+edition templates keep writing ``{% load custom_tags %}`` and get the union of
+both sets. Platform templates load ``platform_tags`` directly and must not load
+``custom_tags`` at all.
+"""
+
+import random
+import re
+from urllib.parse import urlencode
+
+import bleach
+import bleach.css_sanitizer
+import qrcode
+import qrcode.image.svg
+from django import template
+from django.conf import settings
+from django.template.context import RequestContext
+from django.urls import resolve
+from django.urls.exceptions import Resolver404
+from django.utils.html import format_html
+from django.utils.http import url_has_allowed_host_and_scheme
+from django.utils.safestring import SafeData, mark_safe
+
+register = template.Library()
+
+
+def is_active(context: RequestContext, name):
+    """Check if the current view is active."""
+    try:
+        match = resolve(context.request.path)
+        return name == match.view_name
+    except Resolver404:
+        # We can't resolve the request path (e.g. on a 404 page!)
+        # so it can't be active.
+        return False
+
+
+@register.simple_tag(takes_context=True)
+def active_view(context: RequestContext, name):
+    return "active" if is_active(context, name) else ""
+
+
+@register.simple_tag(takes_context=True)
+def active_path(context: RequestContext, *prefixes):
+    """Return 'active' if the current path starts with any of the given prefixes."""
+    try:
+        path = context.request.path
+        return "active" if any(path.startswith(p) for p in prefixes) else ""
+    except (AttributeError, Resolver404):
+        return ""
+
+
+@register.simple_tag(takes_context=True)
+def active_aria(context: RequestContext, name):
+    return 'aria-current="page"' if is_active(context, name) else ""
+
+
+@register.simple_tag(takes_context=True)
+def active_query(context: RequestContext, key, value):
+    return "active" if context["request"].GET.get(key, "") == value else ""
+
+
+@register.simple_tag(takes_context=True)
+def active_query_aria(context: RequestContext, key, value):
+    return 'aria-current="page"' if active_query(context, key, value) else ""
+
+
+@register.simple_tag(takes_context=True)
+def flash(context: RequestContext, id):
+    request = context["request"]
+    return "flash-warn" if request.GET.get("flash") == str(id) else ""
+
+
+@register.filter
+def lookup(dictionary, key):
+    if dictionary is None:
+        return None
+    if isinstance(dictionary, list):
+        # TODO: This assumes the list is a namedtuple with a 'grouper' attribute. This is a bit of a hack.
+        item = next((item for item in dictionary if item.grouper == key), None)
+        return item.list if item else None
+    return dictionary.get(key)
+
+
+@register.filter
+def get_item(obj, key):
+    """Get an item from an object by key using [] access.
+
+    Works with forms (to access fields by name), dicts, and any object with __getitem__.
+    Returns None if the key doesn't exist.
+    """
+    if obj is None:
+        return None
+    try:
+        return obj[key]
+    except (KeyError, TypeError, IndexError):
+        return None
+
+
+@register.filter
+def join_names(items, attr="name"):
+    """Join a list of objects by an attribute, comma-separated.
+
+    Usage: {{ rules|join_names }} or {{ items|join_names:"type" }}
+    """
+    return ", ".join(str(getattr(item, attr, item)) for item in items)
+
+
+@register.simple_tag
+def qt(request, **kwargs):
+    updated = request.GET.copy()
+    for k, v in kwargs.items():
+        if v is not None:
+            updated[k] = v
+        else:
+            updated.pop(k, 0)  # Remove or return 0 - aka, delete safely this key
+
+    return updated.urlencode()
+
+
+def _qt_drop(querydict, drop):
+    """Remove comma-separated keys from a mutable QueryDict copy (no-op if falsy).
+
+    Lets query-string edits invalidate params that no longer apply — e.g. the
+    dice roller drops its reproducible-roll ``seed`` when the dice change.
+    """
+    if not drop:
+        return
+    for key in str(drop).split(","):
+        querydict.pop(key, 0)
+
+
+@register.simple_tag
+def qt_nth(request, drop=None, **kwargs):
+    nth = kwargs.pop("nth")
+    updated = request.GET.copy()
+    _qt_drop(updated, drop)
+    for k, v in kwargs.items():
+        current = updated.getlist(k)
+        if nth < len(current):
+            current[nth] = v
+        else:
+            current.append(v)
+        updated.setlist(k, current)
+
+    return updated.urlencode()
+
+
+@register.simple_tag
+def qt_rm_nth(request, drop=None, **kwargs):
+    nth = kwargs.pop("nth")
+    updated = request.GET.copy()
+    _qt_drop(updated, drop)
+    for k, v in kwargs.items():
+        if str(v) != "1":
+            continue
+        current = updated.getlist(k)
+        if nth < len(current):
+            current.pop(nth)
+            updated.setlist(k, current)
+
+    return updated.urlencode()
+
+
+@register.simple_tag
+def qt_append(request, drop=None, **kwargs):
+    updated = request.GET.copy()
+    _qt_drop(updated, drop)
+    for k, v in kwargs.items():
+        current = updated.getlist(k)
+        current.append(v)
+        updated.setlist(k, current)
+
+    return updated.urlencode()
+
+
+@register.simple_tag
+def qt_rm(request, *args):
+    updated = request.GET.copy()
+    for k in args:
+        updated.pop(k, 0)
+
+    return updated.urlencode()
+
+
+@register.simple_tag
+def qt_getlist(request, key):
+    """
+    Get all values for a query parameter key as a list.
+
+    Use this to properly iterate over multi-value query params (like cat, al).
+    Unlike request.GET.key (which returns only the last value as a string),
+    this uses getlist() to get all values.
+
+    Usage:
+        {% qt_getlist request "cat" as cat_values %}
+        {% for val in cat_values %}<input type="hidden" name="cat" value="{{ val }}">{% endfor %}
+    """
+    return request.GET.getlist(key)
+
+
+@register.simple_tag
+def qt_contains(request, key, value):
+    value = str(value)
+    return key in request.GET and value in request.GET.getlist(key, list)
+
+
+@register.simple_tag
+def qt_has_key(request, key):
+    """Check if a query parameter key exists in the request (even if empty)."""
+    return key in request.GET
+
+
+@register.filter(name="abs")
+def fabs(value):
+    return abs(int(value))
+
+
+@register.filter(name="min")
+def fmin(value, arg):
+    return min(int(value), int(arg))
+
+
+@register.filter(name="max")
+def fmax(value, arg):
+    return max(int(value), int(arg))
+
+
+def identity(value):
+    return value
+
+
+@register.filter
+def to_str(value):
+    """converts int to string"""
+    return str(value)
+
+
+@register.simple_tag
+def qr_svg(value):
+    code = qrcode.make(
+        value, image_factory=qrcode.image.svg.SvgPathImage, box_size=10, border=0
+    ).to_string(encoding="unicode")
+
+    code = re.sub(r'width="\d+mm"', "", code)
+    code = re.sub(r'height="\d+mm"', "", code)
+    code = re.sub(r"<svg ", '<svg width="100%" height="100%" ', code)
+
+    return mark_safe(code)
+
+
+@register.simple_tag
+def settings_value(name):
+    return getattr(settings, name, "")
+
+
+@register.filter
+def subtract(value, arg):
+    """Subtract arg from value."""
+    try:
+        return int(value) - int(arg)
+    except (ValueError, TypeError):
+        return 0
+
+
+@register.simple_tag
+def random_404_message():
+    """Return a random 404 error message."""
+    messages = [
+        "…or maybe that's what the Delaque want you to believe.",
+        "But none can escape Helmawr's justice forever.",
+        "We blame the Caryatids.",
+        "Please direct all queries to your local Archeotek.",
+        "It might have wandered too far into Hive Secundus.",
+        "You probably shouldn't have paid the Whisper Merchant in advance.",
+        "[REDACTED BY ORDER OF LORD HELMAWR]",
+        "Never trust directions from a Goliath.",
+        "You're never going to make it as a Technomancer.",
+        "My cyber-mastiff ate it.",
+        "Check your dome runner map is the right way up.",
+        "Someone might have been at it with a data-thief.",
+        "Guess it wasn't master-crafted after all.",
+        "The engiseer said it was 'working as intended'.",
+        "Nobody knows what happened to it.",
+        "Try adjusting your augmetics.",
+        "Lost in the Ash Wastes without a rebreather.",
+        "Your data-slate was confiscated by Palanite Enforcers.",
+        "Corpse Grinders ate this page. And the server.",
+        "The Redemptionists burned it for heresy.",
+        "Sold to the Mercator Slavers for two credits.",
+        "Currently being digested by a sump kroc.",
+        "Kal Jericho shot first. The page didn't make it.",
+        "Van Saar's tech failed. Again.",
+        "I was trying out some chems and accidentally dissolved the data.",
+        "It fell through a rusted gantry.",
+        "The Genestealers doesn't want you to see this.",
+        "Scragged by a juve on their first raid.",
+        "Your access codes expired 10,000 years ago.",
+        "Currently under quarantine.",
+        "Hunted for sport.",
+        "There was an unscheduled rapid disassembly of the data shaft.",
+        "Lost in transit.",
+        "Eaten by rats. Big ones. With extra eyes.",
+        "The manufactorum quota doesn't include this page.",
+        "Chaos corruption detected. Purge initiated.",
+        "Traded for a rusty stub gun.",
+        "House Cawdor recycled it into a holy flamer.",
+        "The Guilders repo'd it for unpaid tolls.",
+        "A rogue servitor deleted the cache.",
+        "Caught in the crossfire?",
+        "Your dome runner gave you the wrong coordinates.",
+        "Our exovator is still digging for it.",
+        "Currently fermenting in a Goliath still.",
+        "The arbitrator ruled against your access request.",
+        "Lost during the Fall of Hive Secundus.",
+        "Venators claimed the bounty on this page.",
+    ]
+    return random.choice(messages)  # nosec
+
+
+@register.simple_tag
+def cachebuster():
+    """
+    Generate a cachebuster string.
+
+    This is used to force browsers to reload forms even if other params don't change.
+    """
+    return hex(int(random.random() * 1e8))[2:]  # nosec
+
+
+@register.simple_tag
+def dot():
+    return mark_safe("&nbsp;·&nbsp;")
+
+
+# Threshold for applying text-nowrap class to short property values
+NOWRAP_THRESHOLD = 40
+
+
+@register.filter
+def nowrap_class(value, threshold=NOWRAP_THRESHOLD):
+    """
+    Return 'text-nowrap' class if value length is under threshold.
+
+    Use this to prevent short property values from wrapping mid-text
+    while allowing long values to wrap naturally.
+
+    Args:
+        value: The string value to check
+        threshold: Maximum length for nowrap (default: 40 characters)
+
+    Usage:
+        <span class="{{ value|nowrap_class }}">{{ value }}</span>
+        <span class="{{ value|nowrap_class:50 }}">{{ value }}</span>
+    """
+    if value is None:
+        return "text-nowrap"
+    str_value = str(value)
+    return "text-nowrap" if len(str_value) <= threshold else ""
+
+
+@register.simple_tag
+def property_nowrap_class(label, value, threshold=NOWRAP_THRESHOLD):
+    """
+    Return 'text-nowrap' class if combined label+value length is under threshold.
+
+    For asset properties displayed as "Label: Value", this checks if the total
+    display length is short enough to prevent wrapping.
+
+    Args:
+        label: The property label
+        value: The property value
+        threshold: Maximum length for nowrap (default: 40 characters)
+
+    Usage:
+        <span class="{% property_nowrap_class label value %}">{{ label }}: {{ value }}</span>
+    """
+    combined = f"{label}: {value}"
+    return "text-nowrap" if len(combined) <= threshold else ""
+
+
+@register.simple_tag(takes_context=True)
+def safe_referer(context: RequestContext, fallback_url="/"):
+    """
+    Get a safe redirect URL from HTTP_REFERER.
+
+    This tag validates the referer header to prevent open redirect vulnerabilities.
+    If the referer is not safe or missing, returns the fallback URL.
+
+    Args:
+        context: The template context containing the request
+        fallback_url: The URL to use if referer is invalid (default: "/")
+
+    Returns:
+        A safe URL string that can be used for redirects
+    """
+    request = context["request"]
+    referer = request.META.get("HTTP_REFERER", "")
+
+    # Validate the referer URL is safe
+    if referer and url_has_allowed_host_and_scheme(
+        url=referer,
+        allowed_hosts={request.get_host()},
+        require_https=request.is_secure(),
+    ):
+        return referer
+
+    # Fall back to the provided fallback URL
+    return fallback_url
+
+
+@register.simple_tag(takes_context=True)
+def return_url_param(context: RequestContext):
+    """
+    Generate a return_url query parameter for the current page.
+
+    Use this tag when creating links that should return the user to the current page.
+    The tag URL-encodes the current full path (including query string).
+
+    Usage:
+        <a href="{% url 'core:some-view' obj.id %}?{% return_url_param %}">Edit</a>
+
+    Returns:
+        str: Query parameter like "return_url=%2Fsome%2Fpath%2F"
+    """
+    request = context["request"]
+    current_url = request.get_full_path()
+    return urlencode({"return_url": current_url})
+
+
+@register.simple_tag(takes_context=True)
+def return_url_field(context: RequestContext):
+    """
+    Render a hidden form field for return_url.
+
+    Use this tag inside forms to preserve the return_url through form submission.
+    The tag looks for return_url in the template context.
+
+    Usage:
+        <form method="post">
+            {% csrf_token %}
+            {% return_url_field %}
+            ...
+        </form>
+
+    Returns:
+        str: Hidden input element like '<input type="hidden" name="return_url" value="...">'
+    """
+    return_url = context.get("return_url", "")
+    if return_url:
+        return format_html(
+            '<input type="hidden" name="return_url" value="{}">',
+            return_url,
+        )
+    return ""
+
+
+@register.filter
+def plain_text_truncate(value, length=150):
+    """
+    Strip all HTML tags and truncate to a specified length for list view previews.
+
+    This filter is designed for list views where rich text content should be shown
+    as plain text. It:
+    1. Strips all HTML tags (including images, formatting, etc.)
+    2. Removes extra whitespace
+    3. Truncates to the specified length with ellipsis if needed
+
+    Args:
+        value: The HTML string to process
+        length: Maximum length for the output (default: 150 characters)
+
+    Usage:
+        {{ campaign.summary|plain_text_truncate }}
+        {{ campaign.summary|plain_text_truncate:200 }}
+    """
+    if not value:
+        return ""
+
+    # Strip all HTML tags using bleach with an empty tag list
+    text = bleach.clean(value, tags=[], strip=True)
+
+    # Normalize whitespace (collapse multiple spaces/newlines into single spaces)
+    text = " ".join(text.split())
+
+    # Truncate if needed
+    length = int(length)
+    if len(text) > length:
+        # Find the last space before the length limit to avoid cutting words
+        truncated = text[:length]
+        last_space = truncated.rfind(" ")
+        if last_space > length // 2:  # Only use word boundary if it's reasonable
+            truncated = truncated[:last_space]
+        return truncated.rstrip() + "…"
+
+    return text
+
+
+@register.filter
+def safe_rich_text(value):
+    """
+    Sanitize and render rich text content safely.
+
+    This filter sanitizes HTML content to prevent XSS attacks while preserving
+    formatting from TinyMCE. It allows safe HTML tags and attributes commonly
+    used in rich text editing but removes potentially dangerous content like
+    script tags, event handlers, and javascript: URLs.
+
+    Usage:
+        {{ content|safe_rich_text|safe }}
+
+    Note: The |safe filter is still needed after this to render the HTML,
+    but the content has been sanitized first.
+    """
+    if not value:
+        return ""
+
+    # Define allowed tags for rich text content
+    ALLOWED_TAGS = [
+        # Text formatting
+        "p",
+        "br",
+        "span",
+        "div",
+        "strong",
+        "b",
+        "em",
+        "i",
+        "u",
+        "s",
+        "strike",
+        "sub",
+        "sup",
+        "mark",
+        # Headers
+        "h1",
+        "h2",
+        "h3",
+        "h4",
+        "h5",
+        "h6",
+        # Lists
+        "ul",
+        "ol",
+        "li",
+        # Quotes and code
+        "blockquote",
+        "pre",
+        "code",
+        "kbd",
+        "samp",
+        "var",
+        # Links and media
+        "a",
+        "img",
+        # Tables
+        "table",
+        "thead",
+        "tbody",
+        "tfoot",
+        "tr",
+        "th",
+        "td",
+        "caption",
+        "colgroup",
+        "col",
+        # Semantic HTML
+        "article",
+        "section",
+        "nav",
+        "aside",
+        "header",
+        "footer",
+        "address",
+        "time",
+        # Formatting
+        "hr",
+        "abbr",
+        "acronym",
+        "cite",
+        "q",
+        "del",
+        "ins",
+    ]
+
+    # Define allowed attributes for specific tags
+    ALLOWED_ATTRIBUTES = {
+        # Links - only allow href and title, no event handlers
+        "a": ["href", "title", "target", "rel"],
+        # Images - basic attributes only, no event handlers
+        "img": ["src", "alt", "title", "width", "height", "class"],
+        # Tables
+        "table": ["class", "border", "cellpadding", "cellspacing"],
+        "td": ["class", "colspan", "rowspan", "align", "valign"],
+        "th": ["class", "colspan", "rowspan", "align", "valign"],
+        # General styling (but no inline JS)
+        "*": ["class", "id", "style"],
+    }
+
+    # Define allowed CSS properties to prevent CSS-based attacks
+    ALLOWED_STYLES = [
+        "color",
+        "background-color",
+        "font-size",
+        "font-weight",
+        "font-style",
+        "text-decoration",
+        "text-align",
+        "vertical-align",
+        "margin",
+        "margin-top",
+        "margin-bottom",
+        "margin-left",
+        "margin-right",
+        "padding",
+        "padding-top",
+        "padding-bottom",
+        "padding-left",
+        "padding-right",
+        "border",
+        "border-width",
+        "border-color",
+        "border-style",
+        "width",
+        "height",
+        "max-width",
+        "max-height",
+        "min-width",
+        "min-height",
+        "display",
+        "float",
+        "clear",
+        "position",
+        "top",
+        "left",
+        "right",
+        "bottom",
+        "line-height",
+        "white-space",
+        "list-style-type",
+    ]
+
+    # Define allowed URL schemes to prevent javascript: and data: URLs
+    ALLOWED_PROTOCOLS = ["http", "https", "mailto", "tel", "/", "#"]
+
+    # Create CSS sanitizer for style attributes
+    css_sanitizer = bleach.css_sanitizer.CSSSanitizer(
+        allowed_css_properties=ALLOWED_STYLES
+    )
+
+    # Sanitize the HTML content
+    # Use strip=True to remove disallowed tags and their content in one pass
+    cleaned = bleach.clean(
+        value,
+        tags=ALLOWED_TAGS,
+        attributes=ALLOWED_ATTRIBUTES,
+        protocols=ALLOWED_PROTOCOLS,
+        css_sanitizer=css_sanitizer,
+        strip=True,  # Remove disallowed tags and their content
+        strip_comments=True,  # Remove HTML comments
+    )
+
+    return mark_safe(cleaned)
+
+
+@register.filter(name="strip")
+def strip_filter(value):
+    """Trim leading/trailing whitespace from rendered slot content.
+
+    Needed by cotton components whose slot lands inside a plain INLINE element,
+    where the surrounding whitespace is not collapsed away. djlint reflows a
+    one-line call site
+
+        <c-actions.link href="…">Cost</c-actions.link>
+
+    onto three lines, so the slot arrives as "\\n    Cost\\n" and the anchor
+    renders with underlined whitespace either side of the word. `.btn` and
+    `.badge` are immune because Bootstrap makes them inline-block, which strips
+    leading and trailing whitespace inside the box; a bare <a> is not.
+
+    Preserves safety: slot content is already-rendered, escaped HTML, so this
+    re-marks it safe rather than escaping it a second time. A plain str input --
+    a `class` prop carrying user text, say -- is returned unmarked and the engine
+    escapes it as usual. Deliberately NOT registered `is_safe=True`: that flag
+    only propagates safeness, which the isinstance check below already does, so
+    carrying it as well just invited the question of whether it skipped escaping.
+    """
+    if value is None:
+        return ""
+    text = value if isinstance(value, str) else str(value)
+    stripped = text.strip()
+    return mark_safe(stripped) if isinstance(value, SafeData) else stripped
