+++ gyrinx/templates/404.html
@@ -1,5 +1,5 @@
-{% extends "core/layouts/base.html" %}
-{% load custom_tags %}
+{% extends "platform/layouts/base.html" %}
+{% load platform_tags %}
 {% block head_title %}
     Error 404
 {% endblock head_title %}
