+++ gyrinx/templates/account/snippets/menu.html
@@ -1,4 +1,4 @@
-{% load allauth i18n custom_tags %}
+{% load allauth i18n platform_tags %}
 {% firstof list_item_class "" as list_item_class %}
 {% firstof list_a_class "" as list_a_class %}
 {% url 'core:account_home' as account_home_url_ %}
