+++ gyrinx/templates/allauth/layouts/base.html
@@ -1,5 +1,5 @@
 {% extends "platform/layouts/base.html" %}
-{% load i18n custom_tags %}
+{% load i18n platform_tags %}
 {% block body %}
     <div id="content" class="container my-3 my-md-5">
         {% if messages %}
