+++ gyrinx/templates/cotton/messages.html
@@ -31,7 +31,7 @@
 
     `get_item` needs no {% load %}: it is one of two filters django-cotton itself
     injects as a template builtin (with `merge`). Note the project ALSO defines a
-    `get_item` in custom_tags — this resolves to cotton's, since a builtin applies
+    `get_item` in platform_tags — this resolves to cotton's, since a builtin applies
     unless a template loads a library that overrides it. They agree on every dict
     lookup, including a missing key; they differ only when the subject is None,
     where cotton's raises and ours returns None. The subject here is always the
