+++ gyrinx/templates/errors/error.html
@@ -1,4 +1,4 @@
-{% extends "core/layouts/base.html" %}
+{% extends "platform/layouts/base.html" %}
 {% block head_title %}
     Error {{ error_code }}
 {% endblock head_title %}
