+++ gyrinx/templates/platform/includes/account_sidebar.html
@@ -1,4 +1,4 @@
-{% load allauth i18n custom_tags %}
+{% load allauth i18n platform_tags %}
 <nav aria-label="Account settings">
     <ul class="nav nav-pills flex-column gap-1">
         {% url 'core:account_home' as account_home_url_ %}
