+++ n23/core/tests/test_xss_protection.py
@@ -12,7 +12,7 @@ from django.test import Client
 from django.urls import reverse
 
 from n23.core.models import BattleNote, Campaign, List
-from n23.core.templatetags.custom_tags import plain_text_truncate, safe_rich_text
+from gyrinx.site.templatetags.platform_tags import plain_text_truncate, safe_rich_text
 from n23.content.models import ContentHouse
